refactor: change model directories and split move model classes to dedicated files
