+++ src/prediction/models/detector.py
@@ -1,86 +1,14 @@
 import numpy as np
 import torch
-from torch import nn
-from torchvision import models
 
-from src.prediction.models.base_detection_head import BaseDetectionHead
-from src.prediction.models.base_detector import BaseDetector
-from src.prediction.models.base_feature_extractor import BaseFeatureExtractor
+from src.prediction.models.base.base_detection_head import BaseDetectionHead
+from src.prediction.models.base.base_detector import BaseDetector
+from src.prediction.models.base.base_feature_extractor import BaseFeatureExtractor
 from src.prediction.models.postprocessor import Postprocessor
 from src.prediction.models.preprocessor import Preprocessor
-from src.utils.constants import N_BOX_COORDS
+from src.prediction.models.resnet50fe import ResNet50FE
+from src.prediction.models.yolo_v1 import YoloV1
 from src.utils.utils import nms, select_device
-
-
-class ResNet50FE(BaseFeatureExtractor):
-    def __init__(self, grid_size: int, output_feature_channels: int):
-        super().__init__()
-        self.grid_size = grid_size
-        self._output_shape = (output_feature_channels, grid_size, grid_size)
-
-        self.resnet = models.resnet50(weights=models.ResNet50_Weights.IMAGENET1K_V1)
-
-        del self.resnet.avgpool
-        del self.resnet.fc
-
-        self.adaptive_pool = nn.AdaptiveAvgPool2d((grid_size, grid_size))
-
-    @property
-    def output_shape(self) -> tuple[int, int, int]:
-        return self._output_shape
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        x = self.resnet.conv1(x)
-        x = self.resnet.bn1(x)
-        x = self.resnet.relu(x)
-        x = self.resnet.maxpool(x)
-
-        x = self.resnet.layer1(x)
-        x = self.resnet.layer2(x)
-        x = self.resnet.layer3(x)
-        x = self.resnet.layer4(x)
-
-        x = self.adaptive_pool(x)
-        return x
-
-
-class YoloV1(BaseDetectionHead):
-    def __init__(
-        self,
-        input_feature_channels: int,
-        grid_size: int,
-        num_boxes: int,
-        num_classes: int,
-        hidden_size: int,
-        leaky_relu_negative_slope: float,
-    ):
-        super().__init__((input_feature_channels, grid_size, grid_size))
-        self.num_boxes = num_boxes
-        self.num_classes = num_classes
-        self.grid_size = grid_size
-
-        in_features = input_feature_channels * grid_size * grid_size
-
-        self.flatten = nn.Flatten()
-        self.fc1 = nn.Linear(in_features, hidden_size)
-        self.leaky_relu = nn.LeakyReLU(leaky_relu_negative_slope)
-
-        out_features = (
-            grid_size * grid_size * num_boxes * (N_BOX_COORDS + 1 + num_classes)
-        )
-        self.fc2 = nn.Linear(hidden_size, out_features)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        x = self.flatten(x)
-        x = self.leaky_relu(self.fc1(x))
-        x = self.fc2(x)
-        return x.view(
-            -1,
-            self.grid_size,
-            self.grid_size,
-            self.num_boxes,
-            N_BOX_COORDS + 1 + self.num_classes,
-        )
 
 
 class Detector(BaseDetector):
